Modification pour unsigned long long.

+++ Rand48.c
@@ -1,28 +1,31 @@
-long next(long a, long xn, long c, long m){
+unsigned long long next(unsigned long long a, 
+unsigned long long xn, unsigned long long c, unsigned long long m){
     return (a*xn+c)%m;
 }
 
-long* generateur(int nombre, long a, long x0, long c, long m){
+unsigned long long* generateur(int nombre, unsigned long long a, 
+unsigned long long x0, unsigned long long c, unsigned long long m){
     long *res = malloc(sizeof(long)*nombre);
     memset(res, 0, sizeof(long)*nombre);
     
     int i;
-    long xn =x0;
+    unsigned long long xn =x0;
     for(i=0; i<nombre; i++)
         xn = res[i] = next(a, xn, c, m);
 
     return res;
 }
 
-long* rand48(int nombre, long a, long x0, long c){
+unsigned long long* rand48(int nombre, unsigned long long a, 
+unsigned long long x0, unsigned long long c){
     return generateur(nombre, a, x0, c, (1<<48));
 }
 
-long* generateur_java(int nombre){
+unsigned long long* generateur_java(int nombre){
     return rand48(nombre, 25214903917, 0, 11);
 }
 
-void affiche(long* entiers, int nombre){
+void affiche(unsigned long long* entiers, int nombre){
     int i;
     for(i=0; i<nombre; i++){
         printf("%l", entiers[i]);
@@ -32,7 +35,7 @@ void affiche(long* entiers, int nombre){
 int main(){
     // calcul des 10 premier termes à partir de la graine 0 : 
     int nombre = 10;
-    long *alea;
+    unsigned long long *alea;
     affiche(alea=generateur_java(nombre), nombre);
 
     free(alea);
